pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "observix-pro"
version = "0.1.0"
description = "Observix Pro: control plane, agent fleet, and indexer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "httpx>=0.27",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "alembic>=1.13",
  "typer>=0.12",
  "PyYAML>=6.0",
]

[project.scripts]
observix-agent = "observix_agent.cli:app"
observix-control-plane = "observix_control_plane.cli:app"
observix = "observix_cli.main:app"
observix-indexer = "observix_indexer.cli:app"



[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.package-data]
observix_cli = ["config_templates/**/*"]


[tool.setuptools.packages.find]
where = ["src"]
include = ["observix_agent*", "observix_common*", "observix_control_plane*", "observix_indexer*", "observix_cli*"]
